Add -Wl,--no-undefined linker flag only in debug builds
When this flag is added, we need to link against libpython.so to resolve
undefined symbols, which is not good in general. C extensions linked
against libpython.so cannot be imported on a monolithic Python (which
does not have libpython.so). This is a real problem when you want to
redistribute compiled C extensions: if you compile it on RedHat/CentOS,
it won't work on Ubuntu/Debian (the reverse works).